packages/api/src/requirements.txt
attrs==21.4.0
click==8.1.3
distlib==0.3.4
dnspython==2.2.1
email-validator==1.2.1
filelock==3.7.1
Flask==2.1.2
Flask-Cors==3.0.10
Flask-json-schema==0.0.5
flask-mongoengine==1.0.0
Flask-PyMongo==2.3.0
Flask-WTF==1.0.1
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.6.1
MarkupSafe==2.1.1
mongoengine==0.24.1
platformdirs==2.5.2
pymongo==4.1.1
pyrsistent==0.18.1
python-dotenv==0.20.0
six==1.16.0
typing_extensions==4.3.0
virtualenv==20.15.1
Werkzeug==2.1.2
WTForms==3.0.1
zipp==3.8.0
